Added getDurationString to handle durations of time for log messages
Seems as though smallest unit used by DurationFormatUtils is milliseconds, any duration shorter than
a millisecond is considered as zero seconds. This is troublesome for logging the duration of
operations that could take under one millisecond, such as transferring a very small file.
To deal with this, durations in log messages that would be 00.000s are now rounded up to 00.001s.
An exception to this rounding is wait time messages as it is valid for some instances of waiting to
be near-instantaneous, in such instances it can be considered that there is no wait time at all.