Fixed formatting of final total/average times and times which have exactly zero milliseconds
Problem time values not always including three millisecond decimal places was due to the timedelta
class not having a dedicated formatting method.
Previously relied on string casting timedelta objects and truncating them to 11 characters, but this
didn't work for times that ending with exactly 0 milliseconds and for times with over nine hours,
such as the final total time.
Needed to create a timedelta formatting function to handle this.